+++ language/duchain/types/typesystemdata.h
@@ -80,7 +80,7 @@ public:
   uint classSize() const;
 
   /// Returns the complete size of this item in memory, including derived class data and derived class appended list data
-  unsigned short itemSize() const;
+  unsigned int itemSize() const;
 
   /// Expensive
   unsigned int hash() const;
