+++ language/duchain/uses.cpp
@@ -49,7 +49,7 @@ class UsesItem {
     return declaration.hash();
   }
   
-  unsigned short int itemSize() const {
+  unsigned int itemSize() const {
     return dynamicSize();
   }
